[24.11.19] 각종 버그 수정 및 장바구니 구현 완료

+++ src/main/java/com/itwillbs/bookjuk/controller/pay/PayController.java
@@ -117,7 +117,7 @@ public class PayController {
 	    String userEmail = paymentService.getUserEmail(userNum);
 
 	    // userPoint는 user_content 테이블에서 memberNum으로 조회
-	    Long userPoint = paymentService.getUserPoint(userEntity.get().getUserNum());
+	    int userPoint = paymentService.getUserPoint(userEntity.get().getUserNum());
 
 	    // 결제 내역을 조회할 때 memberNum을 사용
 	    List<PaymentEntity> userPayments = paymentService.getPaymentsByMemberNum(userNum);
